requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
python-dotenv==0.19.0
hikerapi==1.0.0
cachetools==4.2.2
Flask-Limiter==2.4.0